Almost done with volunteer and it's UI.
I have a little problem with it though. After a user signs up for a job,
the program terminates. 
This makes it hard to test my code for "view my jobs" because a user
needs to have signed up for a job before they view their jobs.
Does anyone know how to fix this?